Move to errors.py
