+++ jbi/errors.py
@@ -11,3 +11,11 @@ class IgnoreInvalidRequestError(Exception):
 
 class ActionError(Exception):
     """Error occurred during Action handling"""
+
+
+class IncompleteStepError(Exception):
+    """Raised when a step could not complete successfully."""
+
+    def __init__(self, context, *args: object) -> None:
+        super().__init__(*args)
+        self.context = context
